+++ backend/src/graphql/like/like.js
@@ -1,5 +1,5 @@
-const {LikeTC} = require("../models/like/like")
-const requireAuthentication = require("../middleware/jwt/requireAuthentication");
+const {LikeTC} = require("../../models/like/like")
+const requireAuthentication = require("../../middleware/jwt/requireAuthentication");
 
 //***************
 //*** QUERIES ***
